It works!
I moved the Markdown.pl script into the lua directory to make running it
easier. New main function to run the convert function. Also a bunch of
new functions to get the path to the running script so it can figure
out the relative path to the Markdown.pl script.