Cleanup unneeded Lua code
Since the Lua translation backend was originally meant to be used as a
command line tool it still had a lot of the argument parsing and options
code. Since it's not being used like that I could safely delete all of
that. That was one of my TODOs in the README so I checked that off.