New command to convert current buffer
The "MarkdownToHtml" command now converts the markdown file currently in
the buffer to HTML. The new command "NewMarkdownToHtml" and works the
exact same way that the old one did. The current problem is that to edit
the current buffer the VimL command can't take any new lines, so
everything has to be on one line.