Edit README to reflect new project scope
I decided that their wasn't much point to trying to add to the Markdown
elements supported by the plugin because I would just end up recreating
Pandoc, which is not the point of this plugin. Instead of trying to
expand the plugin into something that it's not, I'll keep it as just the
simple John Gruber Markdown.