Update TODO to reflect new project scope
As a stated in my previous commit, I'm keeping this plugin as only a
translator for the John Gruber Markdown syntax. It will only translate
the basic stuff, not all the fancy Pandoc and GitHub style stuff. This
will make it easier to maintain and stay within the scope of the
project. For heavy duty stuff there are better options out there. To
reflect these changes I deleted the todos for adding support for more
Markdown elements.